components/worker_pool/python/requirements.txt
PyYAML==5.2
python-nomad==1.2.1
openstacksdk==0.46.0
